fix(frontend): move bootApp out of App.tsx to restore fast refresh
App.tsx's non-component bootApp export (added in #67564) breaks React
Fast Refresh for the whole app: any HMR invalidation that reaches
App.tsx cascades into src/index.tsx and forces a full page reload.
Under Vite 8 (#68343) an invalidated entry additionally leaves the dev
server serving a double-executing entry module (two createRoot calls on
#root -> removeChild crash on every load) until the server restarts.

Moving bootApp to its own module makes scenes/App component-only again,
so cascades are absorbed there instead of reaching the entry.
